+++ example/app/src/App.vue
@@ -144,6 +144,8 @@
       <li>
         <hr/>
         <button @click="createSubscription">createSubscription</button>
+        <hr/>
+        <button @click="removeSubscription">removeSubscription</button>
       </li>
     </ul>
     <br />
@@ -215,6 +217,10 @@ export default {
   mounted() {
     // Initialize the provider based client
     this.connect = new mewConnect.Provider({windowClosedError: true});
+
+    this.connect.on('popupWindowClosed', () =>{
+      console.log(`popup window closed EVENT`);
+    })
     // this.connect = new mewConnect.Provider();
     // Create the MEWconnect web3 provider
     this.ethereum = this.connect.makeWeb3Provider(1)
@@ -233,6 +239,7 @@ export default {
     this.connect.on('disconnected', () => {
       console.log(`accountsChanged User's address is DISCONNECTED`);
     });
+
 
 
     this.altPopup = new PopUpCreator();
@@ -448,7 +455,7 @@ export default {
       this.web3.eth.getChainId().then(res => (this.chainId = res));
     },
     createSubscription() {
-      let subscription = this.web3.eth
+      this.subscription = this.web3.eth
         .subscribe('newBlockHeaders', function(error, result) {
           if (!error) {
             console.log(result);
@@ -465,6 +472,12 @@ export default {
         .on('connected', function(transaction) {
           console.log(transaction);
         });
+    },
+    removeSubscription() {
+      this.subscription.unsubscribe(function(error, success){
+        if(success)
+          console.log('Successfully unsubscribed!');
+      });
     }
   }
 };